+++ CHANGES.rst
@@ -8,6 +8,7 @@ Changes
 * Add support for Django-5.1.
 * Experimental support for animated image formats. See documentation for more infos.
 * Fix #642: Do not scale images (SVG) without size information.
+* Fix #366: Keep ICC profile when saving image, if present.
 
 
 2.9 (2024-07-25)